Sprint redesign of index layout
Quick recap, I wasn’t sure about having a multitude of pages, so i
tried to condense it into 2 pages, this new index page implements that
with the bio link anchoring at the bottom of the index and only a link
to projects index. The headers have been updated to reflect those
additions.